+++ plugins/falco-agent/pkg/kube/client.go
@@ -1,43 +1,15 @@
-package main
+package kube
 
 import (
-	"context"
-	"fmt"
-	"net/http"
-
-	"github.com/gorilla/mux"
-	"github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/api/meta"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/restmapper"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"github.com/sirupsen/logrus"
 )
 
-const port = "3031"
-
-func main() {
-	r := mux.NewRouter()
-	dynamic, restMapper, err := getKubeClient()
-	if err != nil {
-		panic(err)
-	}
-	r.HandleFunc("/data", func(w http.ResponseWriter, r *http.Request) {
-		inputDataHandler(w, r, context.Background(), dynamic, restMapper)
-	}).Methods(http.MethodPost)
-	r.HandleFunc("/output", outputDataHandler).Methods(http.MethodGet)
-	srv := &http.Server{
-		Handler: r,
-		Addr:    fmt.Sprintf(":%s", port),
-	}
-	r.HandleFunc("/health", func(w http.ResponseWriter, r *http.Request) {
-		w.Write([]byte("OK"))
-	})
-	logrus.Infof("server is running at http://0.0.0.0:%s", port)
-	logrus.Fatal(srv.ListenAndServe())
-}
-
-func getKubeClient() (dynamic.Interface, meta.RESTMapper, error) {
+func GetKubeClient() (dynamic.Interface, meta.RESTMapper, error) {
 	var restMapper meta.RESTMapper
 	var dynamicClient dynamic.Interface
 	kubeConf, configError := ctrl.GetConfig()
@@ -66,3 +38,4 @@ func getKubeClient() (dynamic.Interface, meta.RESTMapper, error) {
 	restMapper = restmapper.NewDiscoveryRESTMapper(resources)
 	return dynamicClient, restMapper, nil
 }
+
